+++ src/components/TransparencyChart/components/IndexChartLegend.tsx
